requirements.txt
Flask
PyJWT
python-dotenv
Werkzeug